+++ types/src/jwks/jwk/mod.rs
@@ -3,24 +3,80 @@
 use crate::{
     jwks::{rsa::RSA_JWK, unsupported::UnsupportedJWK},
     move_any::{Any as MoveAny, AsMoveAny},
+    move_utils::as_move_value::AsMoveValue,
 };
 use anyhow::anyhow;
+use aptos_crypto_derive::{BCSCryptoHash, CryptoHasher};
+use move_core_types::value::{MoveStruct, MoveValue};
 use serde::{Deserialize, Serialize};
+use std::{
+    cmp::Ordering,
+    fmt::{Debug, Formatter},
+};
 
 /// Reflection of Move type `0x1::jwks::JWK`.
 /// When you load an on-chain config that contains some JWK(s), the JWK will be of this type.
 /// When you call a Move function from rust that takes some JWKs as input, pass in JWKs of this type.
 /// Otherwise, it is recommended to convert this to the rust enum `JWK` below for better rust experience.
-#[derive(Debug, PartialEq, Serialize, Deserialize)]
+/// See its doc in Move for more details.
+#[derive(Clone, Eq, PartialEq, Serialize, Deserialize, CryptoHasher, BCSCryptoHash)]
 pub struct JWKMoveStruct {
     pub variant: MoveAny,
 }
 
+impl Debug for JWKMoveStruct {
+    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
+        let jwk = JWK::try_from(self);
+        f.debug_struct("JWKMoveStruct")
+            .field("variant", &jwk)
+            .finish()
+    }
+}
+
+impl AsMoveValue for JWKMoveStruct {
+    fn as_move_value(&self) -> MoveValue {
+        MoveValue::Struct(MoveStruct::Runtime(vec![self.variant.as_move_value()]))
+    }
+}
+
 /// The JWK type that can be converted from/to `JWKMoveStruct` but easier to use in rust.
-#[derive(Debug, PartialEq)]
+#[derive(Debug, PartialEq, Eq)]
 pub enum JWK {
     RSA(RSA_JWK),
     Unsupported(UnsupportedJWK),
+}
+
+impl JWK {
+    pub fn id(&self) -> Vec<u8> {
+        match self {
+            JWK::RSA(rsa) => rsa.id(),
+            JWK::Unsupported(unsupported) => unsupported.id(),
+        }
+    }
+}
+
+impl PartialOrd for JWK {
+    fn partial_cmp(&self, other: &Self) -> Option<Ordering> {
+        Some(self.cmp(other))
+    }
+}
+
+impl Ord for JWK {
+    fn cmp(&self, other: &Self) -> Ordering {
+        self.id().cmp(&other.id())
+    }
+}
+
+impl From<serde_json::Value> for JWK {
+    fn from(value: serde_json::Value) -> Self {
+        match RSA_JWK::try_from(&value) {
+            Ok(rsa) => Self::RSA(rsa),
+            Err(_) => {
+                let unsupported = UnsupportedJWK::from(value);
+                Self::Unsupported(unsupported)
+            },
+        }
+    }
 }
 
 impl From<JWK> for JWKMoveStruct {
@@ -55,46 +111,5 @@ impl TryFrom<&JWKMoveStruct> for JWK {
     }
 }
 
-#[test]
-fn convert_jwk_move_struct_to_jwk() {
-    let unsupported_jwk = UnsupportedJWK::new_for_testing("id1", "payload1");
-    let jwk_move_struct = JWKMoveStruct {
-        variant: unsupported_jwk.as_move_any(),
-    };
-    assert_eq!(
-        JWK::Unsupported(unsupported_jwk),
-        JWK::try_from(&jwk_move_struct).unwrap()
-    );
-
-    let rsa_jwk = RSA_JWK::new_for_testing("kid1", "kty1", "alg1", "e1", "n1");
-    let jwk_move_struct = JWKMoveStruct {
-        variant: rsa_jwk.as_move_any(),
-    };
-    assert_eq!(JWK::RSA(rsa_jwk), JWK::try_from(&jwk_move_struct).unwrap());
-
-    let unknown_jwk_variant = MoveAny {
-        type_name: "type1".to_string(),
-        data: vec![],
-    };
-    assert!(JWK::try_from(&JWKMoveStruct {
-        variant: unknown_jwk_variant
-    })
-    .is_err());
-}
-
-#[test]
-fn convert_jwk_to_jwk_move_struct() {
-    let unsupported_jwk = UnsupportedJWK::new_for_testing("id1", "payload1");
-    let jwk = JWK::Unsupported(unsupported_jwk.clone());
-    let jwk_move_struct = JWKMoveStruct {
-        variant: unsupported_jwk.as_move_any(),
-    };
-    assert_eq!(jwk_move_struct, JWKMoveStruct::from(jwk));
-
-    let rsa_jwk = RSA_JWK::new_for_testing("kid1", "kty1", "alg1", "e1", "n1");
-    let jwk = JWK::RSA(rsa_jwk.clone());
-    let jwk_move_struct = JWKMoveStruct {
-        variant: rsa_jwk.as_move_any(),
-    };
-    assert_eq!(jwk_move_struct, JWKMoveStruct::from(jwk));
-}
+#[cfg(test)]
+mod tests;
